Updating a lot of things

+++ webroot/js/scripts.js
@@ -5,12 +5,12 @@ Dropzone.options.drop = {
             // Create the remove button
             var removeButton = Dropzone.createElement('<button class="btn btn-sm btn-danger"><i class="fa fa-trash-o"></i></button>');
             var viewButton = Dropzone.createElement('<button class="btn btn-sm btn-info pull-right"><i class="fa fa-search"></i></button>');
-            var setCoverButton = Dropzone.createElement('<button class="btn btn-sm btn-info pull-right"><i class="fa fa-picture-o"></i></button>');
 
             var base_url = jQuery("#folderinfo").data("public-folder-path");
 
             var name = file.name;
             var cover = ((file.cover == 'Y') ? true : false);
+            var file_id = file.id;
             var path = base_url + name;
             var th_path = base_url + "TH/" + name;
 
@@ -28,46 +28,31 @@ Dropzone.options.drop = {
             });
 
 
-            setCoverButton.addEventListener("click", function(e) {
-                e.preventDefault();
-                e.stopPropagation();
-
-                var file_id = file.id;
-
-                // Ajax to set cover
-            })
-
-
-            // Listen to the click event
+            // Removing file
             removeButton.addEventListener("click", function (e) {
                 // Make sure the button click doesn't submit the form:
                 e.preventDefault();
                 e.stopPropagation();
 
+                var resp = confirm("Are you sure?");
 
-                var baseuri = jQuery("body").data("base-url");
+                if (resp) {
+                    var baseuri = jQuery("body").data("plugin-base-url");
+
+                    $.ajax({
+                        url: baseuri + "/pictures/delete/" + file_id,
+                        context: document.body
+                    }).done(function () {
+                            // Remove the file preview.
+                            _this.removeFile(file);
+                        });
+                }
 
 
-                $.ajax({
-                    url: baseuri + "pictures/delete/",
-                    context: document.body
-                }).done(function () {
-                        $(this).addClass("done");
-                    });
-
-                // Remove the file preview.
-                _this.removeFile(file);
-                // If you want to the delete the file on the server as well,
-                // you can do the AJAX request here.
             });
 
             // Add the button to the file preview element.
             file.previewElement.appendChild(removeButton);
-
-            if(!cover){
-                file.previewElement.appendChild(setCoverButton);
-            }
-
         });
     }
 };
